探花

+++ tanhua-model/src/main/java/com/itheima/vo/TodayBest.java
@@ -7,13 +7,14 @@ import org.springframework.beans.BeanUtils;
 
 @Data
 public class TodayBest {
-    private Integer id; //用户id
+    private Long id; //用户id
     private String avatar;
     private String nickname;
     private String gender; //性别 man woman
     private Integer age;
     private String[] tags;
-    private Integer fateValue; //缘分值
+    private Long fateValue; //缘分值
+
 
     public static TodayBest init(UserInfo userInfo, RecommendUser recommendUser) {
         TodayBest vo = new TodayBest();
@@ -21,7 +22,7 @@ public class TodayBest {
         if(userInfo.getTags() != null) {
             vo.setTags(userInfo.getTags().split(","));
         }
-        vo.setFateValue(recommendUser.getScore().intValue());
+        vo.setFateValue(recommendUser.getScore().longValue());
         return vo;
     }
 }
